+++ src/utils/utils.h
@@ -2,15 +2,12 @@
 #define UTILS_H
 
 #include <stddef.h>
+#include <unistd.h>
 
 void log_error(const char *message);
-long int calculate_file_size(char *file_path);
-int read_file(char *file_content, char *file_path, long file_size);
+ssize_t calculate_file_size(char *file_path);
+int read_file(char *file_content, char *file_path, size_t file_size);
 int build_absolute_path(char *buffer, const char *path);
 int load_values_from_file(void *structure, const char *path);
-
-/* size_t calculate_combined_strings_length(unsigned int num_strings, ...); */
-
-/* char *retrieve_header(const char *request, const char *key); */
 
 #endif
